+++ pysrc/SBConvolve.cpp
@@ -57,9 +57,33 @@ namespace galsim {
 
     };
 
+    struct PySBAutoConvolve 
+    {
+        static void wrap() {
+            bp::class_< SBAutoConvolve, bp::bases<SBProfile> >("SBAutoConvolve", bp::no_init)
+                .def(bp::init<const SBProfile &>((bp::args("adaptee"))))
+                .def(bp::init<const SBAutoConvolve &>())
+                ;
+        }
+
+    };
+
+    struct PySBAutoCorrelate 
+    {
+        static void wrap() {
+            bp::class_< SBAutoCorrelate, bp::bases<SBProfile> >("SBAutoCorrelate", bp::no_init)
+                .def(bp::init<const SBProfile &>((bp::args("adaptee"))))
+                .def(bp::init<const SBAutoConvolve &>())
+                ;
+        }
+
+    };
+
     void pyExportSBConvolve() 
     {
         PySBConvolve::wrap();
+        PySBAutoConvolve::wrap();
+        PySBAutoCorrelate::wrap();
     }
 
 } // namespace galsim