+++ tests/test_SBProfile.py
@@ -153,11 +153,11 @@ def do_shoot(prof, img, name):
     compar.setFlux(test_flux)
     compar.draw(img, normalization="surface brightness")
     print 'img.sum = ',img.array.sum(),'  cf. ',test_flux/(dx*dx)
-    np.testing.assert_almost_equal(img.array.sum() * dx*dx, test_flux, 5,
+    np.testing.assert_almost_equal(img.array.sum() * dx*dx, test_flux, 4,
             err_msg="Surface brightness normalization for %s disagrees with expected result"%name)
     compar.draw(img, normalization="flux")
     print 'img.sum = ',img.array.sum(),'  cf. ',test_flux
-    np.testing.assert_almost_equal(img.array.sum(), test_flux, 5,
+    np.testing.assert_almost_equal(img.array.sum(), test_flux, 4,
             err_msg="Flux normalization for %s disagrees with expected result"%name)
 
     prof.setFlux(test_flux)
@@ -2246,6 +2246,121 @@ def test_draw():
     t2 = time.time()
     print 'time for %s = %.2f'%(funcname(),t2-t1)
 
+def test_autoconvolve():
+    """Test that auto-convolution works the same as convolution with itself.
+    """
+    import time
+    t1 = time.time()
+
+    mySBP = galsim.SBMoffat(beta=3.8, fwhm=1.3, flux=5)
+    myConv = galsim.SBConvolve([mySBP,mySBP])
+    myImg1 = galsim.ImageF(80,80)
+    myImg1.setScale(0.4)
+    myConv.draw(myImg1.view())
+    myAutoConv = galsim.SBAutoConvolve(mySBP)
+    myImg2 = galsim.ImageF(80,80)
+    myImg2.setScale(0.4)
+    myAutoConv.draw(myImg2.view())
+    printval(myImg1, myImg2)
+    np.testing.assert_array_almost_equal(
+            myImg1.array, myImg2.array, 4,
+            err_msg="Moffat convolved with self disagrees with SBAutoConvolve result")
+
+    # Repeat with the GSObject version of this:
+    psf = galsim.Moffat(beta=3.8, fwhm=1.3, flux=5)
+    conv = galsim.Convolve([psf,psf])
+    conv.draw(myImg1)
+    conv2 = galsim.AutoConvolve(psf)
+    conv2.draw(myImg2)
+    printval(myImg1, myImg2)
+    np.testing.assert_array_almost_equal(
+            myImg1.array, myImg2.array, 4,
+            err_msg="Moffat convolved with self disagrees with AutoConvolve result")
+
+    # For a symmetric profile, AutoCorrelate is the same thing:
+    conv2 = galsim.AutoCorrelate(psf)
+    conv2.draw(myImg2)
+    printval(myImg1, myImg2)
+    np.testing.assert_array_almost_equal(
+            myImg1.array, myImg2.array, 4,
+            err_msg="Moffat convolved with self disagrees with AutoCorrelate result")
+
+    # Test photon shooting.
+    do_shoot(conv2,myImg2,"AutoConvolve(Moffat)")
+
+    # Also check AutoConvolve with an assymetric profile.
+    # (AutoCorrelate with this profile is done below...)
+    obj1 = galsim.Gaussian(sigma=3., flux=4)
+    obj1.applyShift(-0.2, -0.4)
+    obj2 = galsim.Gaussian(sigma=6., flux=1.3)
+    obj2.applyShift(0.3, 0.3)
+    add = galsim.Add(obj1, obj2)
+    conv = galsim.Convolve([add, add])
+    conv.draw(myImg1)
+    corr = galsim.AutoConvolve(add)
+    corr.draw(myImg2)
+    printval(myImg1, myImg2)
+    np.testing.assert_array_almost_equal(
+            myImg1.array, myImg2.array, 4,
+            err_msg="Asymmetric sum of Gaussians convolved with self disagrees with "+
+            "AutoConvolve result")
+
+    t2 = time.time()
+    print 'time for %s = %.2f'%(funcname(),t2-t1)
+
+
+def test_autocorrelate():
+    """Test that auto-correlation works the same as convolution with the mirror image of itself.
+
+    (See the Signal processing Section of http://en.wikipedia.org/wiki/Autocorrelation)
+    """
+    import time
+    t1 = time.time()
+    # Use a function that is NOT two-fold rotationally symmetric, e.g. two different flux Gaussians
+    myGauss1 = galsim.SBGaussian(sigma=3., flux=4)
+    myGauss1.applyShift(-0.2, -0.4)
+    myGauss2 = (galsim.SBGaussian(sigma=6., flux=1.3))
+    myGauss2.applyShift(0.3, 0.3)
+    mySBP1 = galsim.SBAdd([myGauss1, myGauss2])
+    mySBP2 = galsim.SBAdd([myGauss1, myGauss2])
+    # Here we rotate by 180 degrees to create mirror image
+    mySBP2.applyRotation(180. * galsim.degrees)
+    myConv = galsim.SBConvolve([mySBP1, mySBP2])
+    myImg1 = galsim.ImageF(80,80)
+    myImg1.setScale(0.7)
+    myConv.draw(myImg1.view())
+    myAutoCorr = galsim.SBAutoCorrelate(mySBP1)
+    myImg2 = galsim.ImageF(80,80)
+    myImg2.setScale(0.7)
+    myAutoCorr.draw(myImg2.view())
+    printval(myImg1, myImg2)
+    np.testing.assert_array_almost_equal(
+            myImg1.array, myImg2.array, 4,
+            err_msg="Asymmetric sum of Gaussians convolved with mirror of self disagrees with "+
+            "SBAutoCorrelate result")
+
+    # Repeat with the GSObject version of this:
+    obj1 = galsim.Gaussian(sigma=3., flux=4)
+    obj1.applyShift(-0.2, -0.4)
+    obj2 = galsim.Gaussian(sigma=6., flux=1.3)
+    obj2.applyShift(0.3, 0.3)
+    add1 = galsim.Add(obj1, obj2)
+    add2 = (galsim.Add(obj1, obj2)).createRotated(180. * galsim.degrees)
+    conv = galsim.Convolve([add1, add2])
+    conv.draw(myImg1)
+    corr = galsim.AutoCorrelate(add1)
+    corr.draw(myImg2)
+    printval(myImg1, myImg2)
+    np.testing.assert_array_almost_equal(
+            myImg1.array, myImg2.array, 4,
+            err_msg="Asymmetric sum of Gaussians convolved with mirror of self disagrees with "+
+            "AutoCorrelate result")
+
+    # Test photon shooting.
+    do_shoot(corr,myImg2,"AutoCorrelate")
+    t2 = time.time()
+    print 'time for %s = %.2f'%(funcname(),t2-t1)
+
 
 if __name__ == "__main__":
     test_gaussian()
@@ -2279,3 +2394,5 @@ if __name__ == "__main__":
     test_rescale()
     test_sbinterpolatedimage()
     test_draw()
+    test_autoconvolve()
+    test_autocorrelate()